% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binmf.R
\name{predict.binmf}
\alias{predict.binmf}
\title{Make predictions for arbitrary entries in matrix}
\usage{
\method{predict}{binmf}(object, a, b = NULL, ...)
}
\arguments{
\item{object}{An object of class "binmf" as returned by function "binmf".}

\item{a}{Row(s) for which to predict.}

\item{b}{Column(s) for which to predict. If NULL, will make predictions for all columns. Otherwise,
it must be of the same length as "a", and the output will contain the prediction for each combination
of "a" and "b" passed here.}

\item{...}{Extra arguments (not used).}
}
\description{
Make predictions for arbitrary entries in matrix
}
\seealso{
\link{binmf}
}
