% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binmf.R
\name{print.binmf}
\alias{print.binmf}
\title{Get information about binmf object}
\usage{
\method{print}{binmf}(x, ...)
}
\arguments{
\item{x}{An object of class "binmf" as returned by function "binmf".}

\item{...}{Extra arguments (not used).}
}
\description{
Print basic properties of a "binmf" object.
}
