% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binmf.R
\name{summary.binmf}
\alias{summary.binmf}
\title{Get information about binmf object}
\usage{
\method{summary}{binmf}(object, ...)
}
\arguments{
\item{...}{Extra arguments (not used).}

\item{x}{An object of class "binmf" as returned by function "binmf".}
}
\description{
Print basic properties of a "binmf" object (same as `print.binmf` function).
}
\seealso{
\link{print.binmf}
}
